+++ controllers/authControllers.js
@@ -1,9 +1,14 @@
 import bcrypt from "bcryptjs";
 import jwt from "jsonwebtoken";
+import fs from "fs/promises";
+import path from "path";
+import Jimp from "jimp";
+import gravatar from "gravatar";
 
 import User from "../models/User.js";
 import { HttpError, ctrlWrapper } from "../helpers/index.js";
 
+const avatarsPath = path.resolve("public", "avatars");
 const SECRET = process.env.JWT_SECRET;
 
 const register = async (req, res) => {
@@ -13,8 +18,26 @@ const register = async (req, res) => {
     throw HttpError(409, "Email in use");
   }
 
+  // console.log(req);
+  // const {path: oldPath, filename } = req.file;
+  // const img = await Jimp.read(oldPath);
+  // await img
+  //   .autocrop()
+  //   .cover(250, 250, Jimp.HORIZONTAL_ALIGN_CENTER || Jimp.VERTICAL_ALIGN_MIDDLE)
+  //   .writeAsync(oldPath);
+
+  //   const newPath = path.join(avatarsPath, filename);
+  //   await fs.rename(oldPath, newPath);
+
+  // const avatarURL = path.join("avatars", filename);
+
+  const avatarURL = gravatar.url(email);
   const hashPassword = await bcrypt.hash(password, 10);
-  const newUser = await User.create({ ...req.body, password: hashPassword });
+  const newUser = await User.create({
+    ...req.body,
+    password: hashPassword,
+    avatarURL,
+  });
 
   res.status(201).json({
     user: { email: newUser.email, subscription: newUser.subscription },
