+++ models/Contact.js
@@ -19,9 +19,6 @@ const contactSchema = new Schema(
       type: Boolean,
       default: false,
     },
-    avatarURL: {
-      type: String,
-    },
     owner: {
       type: Schema.Types.ObjectId,
       ref: "user",
@@ -56,7 +53,6 @@ export const createContactSchema = Joi.object({
     "string.min": `"phone" should have a minimum length of {#limit}`,
   }),
   favorite: Joi.boolean(),
-  avatarURL: Joi.string(),
 });
 
 export const updateContactSchema = Joi.object({
