+++ package.json
@@ -10,6 +10,7 @@
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
     "express": "^4.18.2",
+    "gravatar": "^1.8.2",
     "jimp": "^0.22.12",
     "joi": "^17.11.0",
     "jsonwebtoken": "^9.0.2",
