+++ routes/contactsRouter.js
@@ -8,7 +8,6 @@ import {
 } from "../models/Contact.js";
 import {
   authenticate,
-  upload,
   validateBody,
   validateId,
 } from "../middlewares/index.js";
@@ -21,7 +20,6 @@ contactsRouter.get("/", contactsControllers.getAllContacts);
 
 contactsRouter.post(
   "/",
-  upload.single("avatar"),
   validateBody(createContactSchema),
   contactsControllers.createContact
 );
